bugfix: added generation stamp to log: we reused the log_t.first_pn as generation stamp field for CREATE and SET_GENSTAMP operations.

+++ hadoop-hdfs/src/main/java/edu/cornell/cs/blog/JNIBlog.java
@@ -133,13 +133,25 @@ public class JNIBlog
    *   CREATE_BLOCK = 1
    *   DELETE_BLOCK = 2
    *   WRITE_BLOCK = 3
+   *   SET_GENSTAMP = 4
+   * genStamp: for SET_GENSTAMP only.
    */
-  private void replayLogOnMetadata(long blockId, int op){
+  private void replayLogOnMetadata(long blockId, int op, long genStamp){
+    if(op == 0) // do nothing for BOL
+      return;
     MemBlockMeta mbm = this.blockMaps.get(blockId);
-    if(mbm == null)
-      mbm = dsmgr.new MemBlockMeta(this,1000l,blockId,ReplicaState.FINALIZED);
-    if(op == 2)
+    if(mbm == null && op == 1){ // CREATE_BLOCK
+      mbm = dsmgr.new MemBlockMeta(this,genStamp,blockId,ReplicaState.FINALIZED);
+    }
+    else if(mbm == null){
+      //create block on an existing block.
+    }else if(op == 2) // WRITE_BLOCK
       mbm.delete();
+    else if(op==3){//WRITE_BLOCK
+      // do nothing
+    }
+    else if(op == 4)
+      mbm.setGenerationStamp(genStamp);
     this.blockMaps.put(blockId, mbm);
   }
   
@@ -149,12 +161,21 @@ public class JNIBlog
   public native void destroy();
   
   /**
+   * set the generationStamp of a blog.
+   * @param mhlc message hybridLogicalClock
+   * @param blockId
+   * @param genStamp
+   * @return
+   */
+  public native int setGenStamp(HybridLogicalClock mhlc, long blockId, long genStamp);
+  
+  /**
    * create a block
    * @param mvc - vector clock of the driven message
    * @param blockId - block identifier
    * @return error code, 0 for success.
    */
-  public native int createBlock(HybridLogicalClock mhlc, long blockId);
+  public native int createBlock(HybridLogicalClock mhlc, long blockId, long genStamp);
   
   /**
    * @param mvc - vector clock of the driven message
@@ -234,17 +255,17 @@ public class JNIBlog
   public void testBlockCreation(HybridLogicalClock mhlc)
   {
     for (long i = 0; i < 100; i++) {
-      assert (createBlock(mhlc,i) == 0);
+      assert (createBlock(mhlc,i,0l) == 0);
       writeLine("Block " + i + " was created.");
     }
     for (long i = 4096; i < 4196; i++) {
-      assert (createBlock(mhlc,i) == 0);
+      assert (createBlock(mhlc,i,0l) == 0);
       writeLine("Block " + i + " was created.");
     }
-    assert (createBlock(mhlc,1) == -1);
-    assert (createBlock(mhlc,4096) == -1);
-    assert (createBlock(mhlc,4100) == -1);
-    assert (createBlock(mhlc,5000) == 0);
+    assert (createBlock(mhlc,1,0l) == -1);
+    assert (createBlock(mhlc,4096,0l) == -1);
+    assert (createBlock(mhlc,4100,0l) == -1);
+    assert (createBlock(mhlc,5000,0l) == 0);
     writeLine("Block 5000 was created.");
   }
